Add timeout option
This is very useful if you need to capture an overlay that disappears
when your mouse moves.

Improvements that could be made.

I did not figure out how to make the short -t to work. Don't see how
they are setup in getopt for the other commands either.

There is no countdown, I only do the send-notification call.